fix memory and swap size calculations
memory.limit_in_bytes limits memory usage, however
memory.memsw.limit_in_bytes limits the *sum* of swap+memory
which means we cannot separately change memory and swap
in update_lxc_config as in order to keep the swap size
unchanged when changing the memory value, the "total" value
has to be recalculated.

Also fixing vmstatus' return value which only returned the
sum in maxmem and left maxswap unset, which strangely
defaults to 512 on the UI side, causing wrong numbers to be
displayed to the user.

Fixes #671