consistent interface names and live network update
veth${vmid}.${id} looks like a vlan device, with qemu we use
tap${vmid}i${id}, so it makes sense to use an 'i' for
containers, too.

Fixed update_net to work with the new configuration method,
it still expected the old configuration hash and errored
when trying to change the network interface configuration
of a running continer.