Rework snapshot config removal logic
Correctly update parent relations in config file upon snapshot removal.

Previously, only the parent of the current state was updated/removed,
which led to broken parent relations if any snapshot other then the
immediate parent of the current snapshot was removed. To fix this,
the parent relation of all children snapshots of the removed snapshot
are updated/removed as well.

Based on code in qemu-server/PVE/QemuServer.pm and parts
of a patch by Gerrit Venema <gmoniker at gmail.com>

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>