Setup::Base: more generic ct_modify_file function
Replaced ct_modify_file_head_portion() with ct_modify_file()
taking a header and and replacing the section in-place
rather than always moving it to the top of the file.

The behavior for when a section is not defined can be
defined via the provided options (replace, prepend, or
append).