automatically restore mountpoints from backup
if the mountpoint configuration is not explicitly provided
via -rootfs or -mpX , recreate all volume mountpoints on the
provided -storage , restore the archive and then add
non-volume mountpoints back to the configuration.

if -rootfs (and optionally any -mpX) are provided, the old
configuration contained in the backup archive is ignored.
this allows reconfiguration of mountpoints when restoring a
backup, e.g. to change mountpoint paths or other options. in
this mode restoring backups to bind or device mountpoints is
also possible.