fix 1153: let crypt() handle unicode CT passwords
according to "perldoc -f crypt", crypt() should downgrade
unicode strings anyway:

If using crypt() on a Unicode string (which potentially has
characters with codepoints above 255), Perl tries to make
sense of the situation by trying to downgrade (a copy of)
the string back to an eight-bit byte string before calling
crypt() (on that copy).  If that works, good.  If not,
crypt() dies with "Wide character in crypt".

login via Spice and ssh works now at least, the noVNC /
vncterm combo seems to be broken because of an unrelated
unicode issue..

Signed-off-by: Fabian Grünbichler <f.gruenbichler@proxmox.com>