setup/suse: use new os_release_var to simplify version detection
Not only simplify but also correct version detection.

Until now we checked the files 'SuSE-release' and 'SuSE-brand' for
version parsing. 'SuSE-release' is marked as obsolete and replaced by
the newer standarised 'os-release', and the fallback is plain wrong
and not guaranteed to exist or match the actual version.

'SuSE-brand' does not get supplied by the 'openSUSE-release' package
but by another package, i.e. 'branding-openSUSE' this isn't
guaranteed to be installed, at least on CT creation, and may have
another version as the actual template provides. E.g. on tumbleweed I
get version 13.3 there, while the release package tells me "20170729".

As "os-release" is available at least sine openSUSE 12.2, and we
support 13.X and newer currently, just use it instead.

Adapt the regex as the non-rolling releases have always a X.Y format.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>