remove transitional pve-update-lxc-config script (PVE 4 beta era)
Short nack history:
In PVE 4 Beta we introduced LXC as our new container technology.
Initially we did not used the our section config format for its
configuration file in /etc/pve . It  was then decided to reuse our
config format (section config), so that we do not need to maintain a
separate parser, and that VM and CT config where not completely
different, which could confuse users.

This script was added to allow an easy transition from the old LXC
config format to the new Proxmox SectionConfig one.

All new installations since, and including, PVE 4.0 never needed this.
And all beta users must go through PVE 4.4 if they want to
dist-upgrade to PVE 5.0, so just remove it - it's forever tracked in
git anyway

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>