fix #2027: do not disable IPV6_AUTOCONF on centos
we used to disable IPV6_AUTOCONF when the DHCP option was chosen for the
container network (was only activated with SLAAC option).

however, this option is actually dependent on IPV6FORWARDING (which is
set to no by default), according to this rule:

IPV6_AUTOCONF=!IPV6FORWARDING

which enables it automatically when forwarding is disabled. this way, we
respect the defaults set by centos.

Signed-off-by: Oguz Bektas <o.bektas@proxmox.com>