CentOS 6: add '--nohangup' as mingetty arg
When creating an unprivileged container with CentOS 6 (which will be EOL in
Nov 2020 [0]) the console does not work.

The problem is mitigated by adding the --nohangup argument to the mingetty
invocations during bootup (in /etc/init/tty.conf).

The idea for the fix is based on the legacy template builder-scripts from
lxc:
https://github.com/lxc/lxc-templates/blob/master/templates/lxc-centos.in#L308

Since '/etc/init/tty.conf' is only written during container creation/restore
and since it is guarded to CentOS versions < 7, the potential for regression
should be rather small.

Tested by creating an unprivileged and a privileged CentOS6 container and
with nesting enabled and disabled for both - the console showed up in
all cases with this fix.

[0] https://wiki.centos.org/About/Product
Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>