setup/debian: use cidr for buster and newer
starting with version 0.8.35 of ifupdown (shipped currently with buster)
the configuration using a separate 'netmask' line instead of providing the
cidr in the 'address' line of a interface stanza of /etc/network/interfaces
is deprecated.

This means that some software installed on newer debian versions, which
parses /etc/network/interfaces may not support the format currently written
by PVE::LXC::Setup::Debian::setup_network.

This patch changes the content of the generated file to use the newer format
only for newer versions of debian (alpine, older ubuntu versions and devuan
also rely on the sub to generate the network config)

caught by installing proxmox-backup-server on a debian buster container and
getting a parse-error in the network configuration tab in the GUI.

tested by creating a ubuntu-14.04, debian-6, debian-8 and a debian-10
container and checking the resulting /etc/network/interfaces.

[0] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=912220

Signed-off-by: Stoiko Ivanov <s.ivanov@proxmox.com>